backend/requirements.txt
# Google ADK
google-adk[all]>=0.1.0

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google Cloud
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
